restructred:restructured the technician folder structure
