+++ src/interfaces/DTO/IServices/Itechnicianservices.dto/technicianAuthService.dto.ts
@@ -1,4 +1,4 @@
-import { Itechnician } from "../../Models/Itechnician";
+import { Itechnician } from "../../../Models/Itechnician";
 
 export interface RegisterResponseDTO {
   success: boolean;
@@ -45,7 +45,6 @@ export interface loginResponseDTO {
   message: string;
   role?: string;
   access_token?: string;
-  refresh_token?: string;
   technician?: Pick<
     Itechnician,
     "username" | "phone" | "email" | "is_verified"
